Make procfs its own package
