+++ procfs/dir.go
@@ -1,0 +1,73 @@
+package procfs
+
+import (
+	"sync"
+	"time"
+
+	db "sigmaos/debug"
+	"sigmaos/fs"
+	"sigmaos/path"
+	"sigmaos/proc"
+	"sigmaos/serr"
+	sp "sigmaos/sigmap"
+)
+
+type ProcDir struct {
+	sync.Mutex
+	*ProcInode
+	procs ProcFs
+	ps    []*proc.Proc
+}
+
+func NewProcDir(procs ProcFs) fs.Inode {
+	return &ProcDir{ProcInode: newProcInode(sp.DMDIR|0444, "pids"), procs: procs}
+}
+
+func (pd *ProcDir) Open(ctx fs.CtxI, m sp.Tmode) (fs.FsObj, *serr.Err) {
+	pd.ps = pd.procs.GetProcs()
+	return nil, nil
+}
+
+func (pd *ProcDir) Close(ctx fs.CtxI, mode sp.Tmode) *serr.Err {
+	pd.ps = nil
+	return nil
+}
+
+func (pd *ProcDir) Create(ctx fs.CtxI, name string, perm sp.Tperm, m sp.Tmode, lid sp.TleaseId, f sp.Tfence) (fs.FsObj, *serr.Err) {
+	return nil, serr.NewErr(serr.TErrNotSupported, "Create")
+}
+
+func (pd *ProcDir) Remove(ctx fs.CtxI, name string, f sp.Tfence) *serr.Err {
+	return serr.NewErr(serr.TErrNotSupported, "Remove")
+}
+
+func (pd *ProcDir) Rename(ctx fs.CtxI, from, to string, f sp.Tfence) *serr.Err {
+	return serr.NewErr(serr.TErrNotSupported, "Rename")
+}
+
+func (pd *ProcDir) Renameat(ctx fs.CtxI, from string, dd fs.Dir, to string, f sp.Tfence) *serr.Err {
+	return serr.NewErr(serr.TErrNotSupported, "Renameat")
+}
+
+func (pd *ProcDir) Stat(ctx fs.CtxI) (*sp.Stat, *serr.Err) {
+	st := sp.NewStat(sp.NewQid(sp.QTDIR, 0, pd.path), pd.perm, 0, pd.name, "schedd")
+	st.Length = uint64(len(pd.procs.GetProcs()))
+	st.Mtime = uint32(time.Now().Unix())
+	return st, nil
+}
+
+func (pd *ProcDir) LookupPath(ctx fs.CtxI, path path.Path) ([]fs.FsObj, fs.FsObj, path.Path, *serr.Err) {
+	name := path[0]
+	db.DPrintf(db.PROCFS, "%v: Lookup %v %v\n", ctx, pd, name)
+	if _, ok := pd.procs.Lookup(name); ok {
+		var o fs.FsObj
+		o = newProcInode(0444, name)
+		return []fs.FsObj{o}, o, path[1:], nil
+	} else {
+		return nil, nil, path, serr.NewErr(serr.TErrNotfound, name)
+	}
+}
+
+func (pd *ProcDir) ReadDir(ctx fs.CtxI, cursor int, cnt sp.Tsize) ([]*sp.Stat, *serr.Err) {
+	return nil, nil
+}
