+++ procmgr/state.go
@@ -26,6 +26,24 @@ func NewProcState() *ProcState {
 	}
 }
 
+func (ps *ProcState) GetProcs() []*proc.Proc {
+	ps.Lock()
+	defer ps.Unlock()
+
+	procs := make([]*proc.Proc, 0, len(ps.spawned))
+	for _, p := range ps.spawned {
+		procs = append(procs, p)
+	}
+	return procs
+}
+
+func (ps *ProcState) Lookup(n string) (*proc.Proc, bool) {
+	if p, ok := ps.spawned[sp.Tpid(n)]; ok {
+		return p, ok
+	}
+	return nil, false
+}
+
 func (ps *ProcState) spawn(p *proc.Proc) {
 	ps.Lock()
 	defer ps.Unlock()
